Criado conversor número string controlado; Modificados tipos de byte para ushort para funcionarem melhor na View.

+++ BancoDados/ModeloXML/Enums.cs
@@ -1,6 +1,6 @@
 ﻿namespace NFeFacil.ModeloXML
 {
-    public enum TiposDocumento : byte
+    public enum TiposDocumento : ushort
     {
         CPF,
         CNPJ,
