+++ BancoDados/ModeloXML/PartesProcesso/PartesNFe/PartesDetalhes/Identificacao.cs
@@ -30,7 +30,7 @@ namespace NFeFacil.ModeloXML.PartesProcesso.PartesNFe.PartesDetalhes
         }
 
         [XmlElement(ElementName = "cUF", Order = 0)]
-        public byte CódigoUF { get; set; }
+        public ushort CódigoUF { get; set; }
 
         [XmlElement(ElementName = "cNF", Order = 1)]
         public int ChaveNF { get; set; }
@@ -39,10 +39,10 @@ namespace NFeFacil.ModeloXML.PartesProcesso.PartesNFe.PartesDetalhes
         public string NaturezaDaOperação { get; set; }
 
         [XmlElement(ElementName = "indPag", Order = 3)]
-        public byte FormaPagamento { get; set; } = 0;
+        public ushort FormaPagamento { get; set; } = 0;
 
         [XmlElement(ElementName = "mod", Order = 4)]
-        public byte Modelo { get; set; } = 55;
+        public ushort Modelo { get; set; } = 55;
 
         [XmlElement(ElementName = "serie", Order = 5)]
         public ushort Serie { get; set; } = 1;
@@ -57,37 +57,37 @@ namespace NFeFacil.ModeloXML.PartesProcesso.PartesNFe.PartesDetalhes
         public string DataHoraSaídaEntrada { get; set; }
 
         [XmlElement(ElementName = "tpNF", Order = 9)]
-        public byte TipoOperação { get; set; } = 1;
+        public ushort TipoOperação { get; set; } = 1;
 
         [XmlElement(ElementName = "idDest", Order = 10)]
-        public byte IdentificadorDestino { get; set; } = 1;
+        public ushort IdentificadorDestino { get; set; } = 1;
 
         [XmlElement(ElementName = "cMunFG", Order = 11)]
         public int CodigoMunicípio { get; set; }
 
         [XmlElement(ElementName = "tpImp", Order = 12)]
-        public byte TipoImpressão { get; set; } = 1;
+        public ushort TipoImpressão { get; set; } = 1;
 
         [XmlElement(ElementName = "tpEmis", Order = 13)]
-        public byte TipoEmissão { get; set; } = 1;
+        public ushort TipoEmissão { get; set; } = 1;
 
         [XmlElement(ElementName = "cDV", Order = 14)]
         public int DígitoVerificador { get; set; }
 
         [XmlElement(ElementName = "tpAmb", Order = 15)]
-        public byte TipoAmbiente { get; set; } = 1;
+        public ushort TipoAmbiente { get; set; } = 1;
 
         [XmlElement(ElementName = "finNFe", Order = 16)]
-        public byte FinalidadeEmissão { get; set; } = 1;
+        public ushort FinalidadeEmissão { get; set; } = 1;
 
         [XmlElement(ElementName = "indFinal", Order = 17)]
-        public byte OperaçãoConsumidorFinal { get; set; } = 1;
+        public ushort OperaçãoConsumidorFinal { get; set; } = 1;
 
         [XmlElement(ElementName = "indPres", Order = 18)]
-        public byte IndicadorPresença { get; set; } = 1;
+        public ushort IndicadorPresença { get; set; } = 1;
 
         [XmlElement(ElementName = "procEmi", Order = 19)]
-        public byte ProcessoEmissão { get; set; } = 0;
+        public ushort ProcessoEmissão { get; set; } = 0;
 
         [XmlElement(ElementName = "verProc", Order = 20)]
         public string VersaoAplicativo { get; set; }
